- Renamed dataManager.py to data_manager.py - Renamed dataScraper.py to data_scraper.py - Imported: preprocess_data, verify_data_classifies and FileManager from classifier.py to data_manager.py in order to accommodate the below changes - Moved validate_datasets() to data_manager.py - Moved check_too_few_columns() to data_manager.py - Moved check_too_few_rows() to data_manager.py - Moved check_file_too_large() to data_manager.py - Moved check_more_than_one_element() to data_manager.py - Moved move_invalid_datasets() to data_manager.py - Moved write_failed_files() to data_manager.py
